Minor rewording
After spending time on the Neo4J community boards, I've noticed that a lot of newbies don't realize you can have multiple Labels on a node.

In addition, newbies don't seem to realize that Labels are implement as a set of nodes with the same Label, speeding up access, which is why labels are so important.

My proposed rewording is an attempt to emphasize this.  The proposed text might be repeating itself too much.  On the other hand, I think it's worth re-emphasizing.